+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+
+- Make `DeferredEntity::new` and `DeferredChanges` public.
+
 ## [0.34.3] - 2025-07-01
 
 ### Fixed